Move CreateCustomerReport and CreateWorkerReport into Reports folder
